Refactored state change tests into single StateChangeEvents test suite
